+++ src/main/webapp/WEB-INF/views/welcome.jsp
@@ -2,6 +2,7 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -13,15 +14,23 @@
         background: #111111;
         position: relative;
     }
-    .imgbox{
+
+    .imgbox {
         position: absolute;
         top: 300px;
-        left:700px;
+        left: 700px;
     }
 </style>
+
 <body>
     <div class="imgbox">
-    <a href="/horror/main"><img src="http://file3.instiz.net/data/file3/2018/03/09/b/a/2/ba21066df826cf82a3c221cf6c8d8b06.gif"></a>
-</div>
+        <a href="/login">
+            <img src="http://file3.instiz.net/data/file3/2018/03/09/b/a/2/ba21066df826cf82a3c221cf6c8d8b06.gif">
+        </a>
+    </div>
 </body>
+<script>
+   
+</script>
+
 </html>
